+++ src/extension.ts
@@ -15,6 +15,9 @@ import getRepoInfo from "./utils/vscode/getRepoInfo";
 import getUserEmail from "./utils/getUserEmail";
 import searchType from "./utils/analytics/searchType";
 import getPRsToPaintPerSHAs from "./utils/vscode/getPRsToPaintPerSHAs";
+import slackhelp from "./utils/analytics/slackhelp";
+import getBlameAuthors from "./utils/getBlameAuthors";
+import explainCode from "./utils/openai/explainCode";
 import createDocs from "./utils/analytics/createDocs";
 
 // repo information
@@ -25,6 +28,10 @@ let userEmail: string | undefined = "";
 let localUser: string | undefined = "";
 // selected shas
 let arrayOfSHAs: string[] = [];
+// Selected block of code
+let selectedBlockOfCode: string | undefined = "";
+// codeExplanation
+let codeExplanation: string | undefined = "";
 
 let octokit: any;
 
@@ -101,6 +108,15 @@ export async function activate(context: vscode.ExtensionContext) {
   octokit = await credentials.getOctokit();
 
   vscode.window.onDidChangeTextEditorSelection(async (selection) => {
+    // // Get text of selected piece of code
+    let selectedCode = "";
+    if (selection.selections.length > 0) {
+      let selectedText = selection;
+      selectedCode= selectedText.textEditor.document.getText(selectedText.selections[0]);
+    }
+    // Replace newlines with \n
+    selectedBlockOfCode = selectedCode.replace(/(\r\n|\n|\r)/gm,"");
+    
     arrayOfSHAs = await getSHAArray(
       selection.selections[0].start.line,
       selection.selections[0].end.line,
@@ -152,6 +168,14 @@ class watermelonSidebar implements vscode.WebviewViewProvider {
           this.sendMessage({
             command: "loading",
           });
+
+          // Call Open AI's API to explain the code
+          codeExplanation = await explainCode({  
+            wrangledBlockOfCode: selectedBlockOfCode as string
+          }).then((res) => {
+            return res.data;
+          });
+
           userEmail = await getUserEmail({ octokit });
           localUser = await getLocalUser();
           if (!arrayOfSHAs.length) {
@@ -228,6 +252,7 @@ class watermelonSidebar implements vscode.WebviewViewProvider {
   }
   public sendMessage(message: any) {
     if (this._view) {
+      message.explanation = codeExplanation;
       this._view.show?.(true); // `show` is not implemented in 1.49 but is for 1.50 insiders
       this._view.webview.postMessage(message);
     }
@@ -280,7 +305,7 @@ class watermelonSidebar implements vscode.WebviewViewProvider {
         lightLogo,
         nonce,
         scriptUri,
-        message.author
+        message.author,
       );
     } else {
       return getInitialHTML(